Refactor project into domain, application, and infrastructure layers
