--- src/app/__init__.py
@@ -1,1 +1,0 @@
-"""Application subpackage for GUI controller and related components."""
