+++ src/application/world_service.py
@@ -1,115 +1,79 @@
-#!/usr/bin/env python3
-"""
-Application controller for HexiRules.
-Separates application/world logic from the Tkinter view.
-"""
+from __future__ import annotations
 
-from typing import Any, Dict, List, Optional, Tuple, Iterable
-import json
-import os
-from hex_rules import HexAutomaton
+from pathlib import Path
+from typing import Dict, Iterable, List, Optional
+
+from typing import cast
+
+from domain.hexidirect.rule_engine import HexAutomaton
+from domain.worlds.world import World
+from infrastructure.persistence.json_world_repository import JsonWorldRepository
+from domain.worlds.repository import WorldRepository
 
 
-class HexiController:
+class WorldService:
     """Encapsulates world management and rule application logic."""
 
-    def __init__(self) -> None:
-        self.worlds: Dict[str, Dict[str, Any]] = {}
+    def __init__(self, repository: Optional[WorldRepository] = None) -> None:
+        self.repository: WorldRepository = repository or JsonWorldRepository()
+        self.worlds: Dict[str, World] = {}
         self.current_world: Optional[str] = None
 
     # Worlds
     def create_world(
         self, name: str, radius: int, _is_hex: bool, rules_text: str
     ) -> None:
-        # _is_hex is ignored for backward compatibility; HexiDirect-only now.
-        world: Dict[str, Any] = {
-            "name": name,
-            "radius": radius,
-            "rules_text": rules_text,
-            "hex": HexAutomaton(radius=radius),
-        }
-        if rules_text:
-            rules = [
-                r.strip()
-                for r in rules_text.replace(";", "\n").split("\n")
-                if r.strip()
-            ]
-            world["hex"].set_rules(rules)
+        world = World(name=name, radius=radius, rules_text=rules_text)
         self.worlds[name] = world
 
-    def select_world(self, name: str) -> Dict[str, Any]:
+    def select_world(self, name: str) -> World:
         if name not in self.worlds:
             raise KeyError(f"Unknown world: {name}")
         self.current_world = name
         return self.worlds[name]
 
-    def get_current_world(self) -> Dict[str, Any]:
+    def get_current_world(self) -> World:
         if not self.current_world or self.current_world not in self.worlds:
             raise RuntimeError("No current world selected")
         return self.worlds[self.current_world]
 
     def current_automaton(self) -> HexAutomaton:
-        w = self.get_current_world()
-        return w["hex"]
+        return self.get_current_world().hex
 
     # Persistence
     def save_world_to_file(
         self, path: str, is_hexidirect: bool, rules_text: str
     ) -> None:
         world = self.get_current_world()
-        world["rules_text"] = rules_text
-        data: Dict[str, Any] = {
-            "name": world["name"],
-            "radius": world["radius"],
-            "is_hex": True,  # kept for backward compatibility
-            "rules_text": world["rules_text"],
-            "hex_cells": [
-                {"q": q, "r": r, "s": cell.state, "d": cell.direction}
-                for (q, r), cell in world["hex"].grid.items()
-                if cell.state != "_"
-            ],
-        }
-        with open(path, "w", encoding="utf-8") as f:
-            json.dump(data, f, indent=2)
+        world.rules_text = rules_text
+        self.repository.save(world, Path(path))
 
     def load_world_from_file(self, path: str) -> str:
-        with open(path, "r", encoding="utf-8") as f:
-            data = json.load(f)
-        name = data.get("name") or os.path.splitext(os.path.basename(path))[0]
-        radius = int(data.get("radius", 8))
-        # Accept old files; ignore Conway fields
-        rules_text = str(data.get("rules_text", ""))
-        self.create_world(name, radius, True, rules_text)
-        world = self.worlds[name]
-        # populate
-        world["hex"].clear()
-        for item in data.get("hex_cells", []):
-            world["hex"].set_cell(
-                int(item["q"]), int(item["r"]), str(item["s"]), item.get("d")
-            )
+        world = cast(World, self.repository.load(Path(path)))
+        name = cast(str, world.name)
+        self.worlds[name] = world
         self.select_world(name)
         return name
 
     # Editing and execution
     def clear(self) -> None:
-        w = self.get_current_world()
-        w["hex"].clear()
+        self.get_current_world().hex.clear()
 
     def randomize(self, states: Iterable[str], p: float = 0.3) -> None:
         import random
 
         w = self.get_current_world()
-        R = int(w.get("radius", 8))
+        R = int(w.radius)
         pool = [s for s in states if s != "_"]
         for q in range(-R // 2, R // 2 + 1):
             for r in range(-R // 2, R // 2 + 1):
                 if abs(q + r) <= R // 2 and random.random() < p:
                     state = random.choice(pool) if pool else "a"
                     direction = random.choice([None, 1, 2, 3, 4, 5, 6])
-                    w["hex"].set_cell(q, r, state, direction)
+                    w.hex.set_cell(q, r, state, direction)
 
     def step(self, rules_text: str) -> List[str]:
-        """Apply a single simulation step and return log messages to display."""
+        """Apply a single simulation step and return log messages."""
         logs: List[str] = []
         if not rules_text:
             return logs
@@ -117,21 +81,19 @@ class HexiController:
         logs.append("STEP: Starting new simulation step")
 
         w = self.get_current_world()
-        # Parse rules
         rules: List[str] = []
         for line in rules_text.split("\n"):
             rules.extend([r.strip() for r in line.split(",") if r.strip()])
         if rules:
-            # Persist
-            w["rules_text"] = "\n".join(rules_text.split("\n"))
+            w.rules_text = "\n".join(rules_text.split("\n"))
             logs.append(f"Rules: {rules}")
-            w["hex"].set_rules(rules)
+            w.hex.set_rules(rules)
             logs.append("Expanded rules:")
-            for i, rule in enumerate(w["hex"].rules, 1):
+            for i, rule in enumerate(w.hex.rules, 1):
                 logs.append(f"  {i}: {rule.rule_str}")
             active_cells = [
                 f"({q},{r}):{cell}"
-                for (q, r), cell in w["hex"].grid.items()
+                for (q, r), cell in w.hex.grid.items()
                 if cell.state != "_"
             ]
             logs.append(f"Active cells before step: {len(active_cells)}")
@@ -139,10 +101,9 @@ class HexiController:
                 logs.append(f"  {cell_info}")
             if len(active_cells) > 10:
                 logs.append(f"  ... and {len(active_cells) - 10} more")
-            # Analyze rule applications (summary only)
             checked_count = 0
             match_count = 0
-            hex_world = w["hex"]
+            hex_world = w.hex
             prev_active_set = {
                 pos for pos, cell in hex_world.grid.items() if cell.state != "_"
             }
@@ -159,10 +120,10 @@ class HexiController:
             logs.append(
                 "Note: When multiple rules from same macro match, one is chosen randomly"
             )
-            w["hex"].step()
+            w.hex.step()
             new_active = [
                 f"({q},{r}):{cell}"
-                for (q, r), cell in w["hex"].grid.items()
+                for (q, r), cell in w.hex.grid.items()
                 if cell.state != "_"
             ]
             logs.append(f"Active cells after step: {len(new_active)}")
@@ -170,9 +131,8 @@ class HexiController:
                 logs.append(f"  {cell_info}")
             if len(new_active) > 10:
                 logs.append(f"  ... and {len(new_active) - 10} more")
-            # Delta summary (HexiDirect)
             new_active_set = {
-                pos for pos, cell in w["hex"].grid.items() if cell.state != "_"
+                pos for pos, cell in w.hex.grid.items() if cell.state != "_"
             }
             births = new_active_set - prev_active_set
             survivals = new_active_set & prev_active_set
@@ -198,5 +158,4 @@ class HexiController:
 
     # Utilities
     def active_count(self) -> int:
-        w = self.get_current_world()
-        return len(w["hex"].get_active_cells())
+        return len(self.get_current_world().hex.get_active_cells())
