+++ src/infrastructure/persistence/json_world_repository.py
@@ -1,0 +1,37 @@
+import json
+from pathlib import Path
+
+from domain.worlds.world import World
+from domain.worlds.repository import WorldRepository
+
+
+class JsonWorldRepository(WorldRepository):
+    """File-based JSON repository for worlds."""
+
+    def save(self, world: World, path: Path) -> None:
+        data = {
+            "name": world.name,
+            "radius": world.radius,
+            "rules_text": world.rules_text,
+            "hex_cells": [
+                {"q": q, "r": r, "s": cell.state, "d": cell.direction}
+                for (q, r), cell in world.hex.grid.items()
+                if cell.state != "_"
+            ],
+        }
+        with path.open("w", encoding="utf-8") as f:
+            json.dump(data, f, indent=2)
+
+    def load(self, path: Path) -> World:
+        with path.open("r", encoding="utf-8") as f:
+            data = json.load(f)
+        name = data.get("name") or path.stem
+        radius = int(data.get("radius", 8))
+        rules_text = str(data.get("rules_text", ""))
+        world = World(name=name, radius=radius, rules_text=rules_text)
+        world.hex.clear()
+        for item in data.get("hex_cells", []):
+            world.hex.set_cell(
+                int(item["q"]), int(item["r"]), str(item["s"]), item.get("d")
+            )
+        return world
